requirements.txt
transformers
openai
faiss-cpu